integer -> bigint -> numeric

+++ data/migrations/20190212230830_students.js
@@ -49,7 +49,7 @@ function createScoresTable(knex) {
         scores.increments();
         scores.text("task").notNullable();
         scores.integer("points").notNullable();
-        scores.bigint("date").notNullable(); // production ONLY
+        scores.numeric("date").notNullable(); // production ONLY
         // scores.integer("date").notNullable(); development ONLY
 
         scores
